No major changes made, but need to write this down, see description
On the inventory select page '/'
I'm trying to set it up such that hovering over and element in the
inventory hints that clicking will remove the element form inventory and
implement click in inventory to remove. Currently, when I register and
event listener only the most recently added element works. This is true
of both "addeventlistener("click"" and of "onclick=" withen the div. I'm
not sure why.